test 10 ok

+++ DataLayer/Product.cs
@@ -9,7 +9,7 @@ public class Product
 {
     public int Id { get; set; } = 0;
     public string Name { get; set; } = null;
-    public double UnitPrice { get; set; } = 0.0;
+    public decimal UnitPrice { get; set; } = 0.0m;
     public int UnitsInStock { get; set; } = 0;
     public string QuantityPerUnit { get; set; } = null;
     public int CategoryId { get; set; }
